backend/requirements.txt
fastapi==0.105.0
uvicorn==0.24.0
pydantic==2.5.2
pydantic-settings==2.1.0
python-dotenv==1.0.0
httpx==0.25.2
beautifulsoup4==4.12.2
lxml==4.9.3
google-generativeai==0.8.4
python-multipart==0.0.6
starlette>=0.27.0
requests>=2.28.0

# データベース関連
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.1

# AI関連
anthropic==0.49.0

# スクリーンショット関連
playwright==1.51.0
pillow==10.2.0